chore: updated ts-unused-exports config for the components workspace
