+++ packages/components/src/index.ts
@@ -1,5 +1,7 @@
 export { Button } from '@/button/Button';
 export { PageMenu } from '@/page-menu/PageMenu';
+export { PageMenuItem } from '@/page-menu/PageMenuItem';
+export { PageSwitcher } from '@/page-switcher/PageSwitcher';
 export { Scrollbar } from '@/scrollbar/Scrollbar';
 export { SelectBox } from '@/select-box/SelectBox';
 export { Slider } from '@/slider/Slider';
