+++ packages/components/src/page-menu/PageMenuItem.model.ts
@@ -1,0 +1,8 @@
+import { CSSProperties, ReactNode } from 'react';
+
+export type PageMouseItemProps = {
+  style?: CSSProperties;
+  children?: ReactNode;
+  onMouseEnter?: any;
+  onMouseLeave?: any;
+};
