+++ packages/components/src/select-box/SelectBoxInput.model.ts
@@ -1,0 +1,5 @@
+export type SelectBoxInputProps = {
+  label: string;
+  placeholder: string;
+  onClick: (value: string) => void;
+};
